backend_flask/requirement.txt
Flask
Flask-CORS
pandas
numpy
scikit-learn
joblib
spacy
requests
python-dotenv
fuzzywuzzy
python-Levenshtein
Pillow
pytesseract
PyPDF2
transformers
torch
opencv-python
easyocr
werkzeug